Implementing a working microphone streaming example for Stream input extension for comparison
